hola

+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: mhumbert <mhumbert@learner.42.tech>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/07 14:44:19 by mhumbert          #+#    #+#             */
-/*   Updated: 2025/11/11 13:39:35 by mhumbert         ###   ########.fr       */
+/*   Updated: 2025/11/11 14:30:14 by mhumbert         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (n == 0)
